Add support for token passwords
If F5VPN_USE_TOKEN is set to a non-empty value
then f5vpn-login will ask for a domain password
and a token password.  This is probably a really horrible
hack but I'm not a python programmer.
I changed the execve
syscall in f5vpn-login-runner.c to execv to forward the current
environment rather than adding the logic to look for and set
the F5VPN_USE_TOKEN variable.